requirements.txt
anyascii~=0.3.1
appdirs~=1.4.4
asgiref~=3.6.0
beautifulsoup4~=4.11.1
boto3~=1.26.54
botocore~=1.29.54
certifi~=2022.12.7
charset-normalizer~=3.0.1
dj-database-url~=1.2.0
dj-static~=0.0.6
Django~=4.1.5
django-appconf~=1.0.5
django-compat~=1.0.15
django-compressor~=4.3
django-el-pagination~=3.3.0
django-filter~=22.1
django-generic-positions~=0.2.3
django-location-field~=2.1.0
django-modelcluster~=6.0
django-permissionedforms~=0.1
django-photologue~=3.15.1
django-social-share~=2.3.0
django-sortedm2m~=3.1.1
django-storages~=1.13.2
django-taggit~=3.1.0
django-toolbelt~=0.0.1
django-treebeard~=4.6.0
djangorestframework~=3.14.0
docutils~=0.19
draftjs-exporter~=2.1.7
et-xmlfile~=1.1.0
ExifRead~=3.0.0
futures~=3.0.5
gunicorn~=20.1.0
html5lib~=1.1
idna~=3.4
jmespath~=1.0.1
l18n~=2021.3
load-dotenv~=0.1.0
marshmallow~=3.19.0
olefile~=0.46
openpyxl~=3.0.10
packaging~=23.0
Pillow~=9.4.0
psycopg2~=2.9.5
pyparsing~=3.0.9
python-dateutil~=2.8.2
python-dotenv~=0.21.1
pytz~=2022.7.1
PyYAML~=6.0
rcssmin~=1.1.1
requests~=2.28.2
rjsmin~=1.2.1
s3transfer~=0.6.0
six~=1.16.0
soupsieve~=2.3.2.post1
sqlparse~=0.4.3
static3~=0.7.0
telepath~=0.3
Unidecode~=1.3.6
urllib3~=1.26.14
wagtail~=4.1.1
webencodings~=0.5.1
whitenoise~=6.3.0
Willow~=1.4.1
